Spotify Music search engine
This program uses Html, css, javascript and jgrasp to search an artist
and it info on spotfy. Ajax request is to spotify to retrive related
artists to the name the user give. Each artist's name and image will be
displayed.  When the user selects an artist from the search list it will
send another ajax request to retrive the authos Info, three albums in
us, five realted artists to him/her... in JSON format whic intrun will
expand itself and displays this info.